use new delimiters for inlined MLE modules
introduced in 23.4, not a string anymore.
Looks like a string in this case, but it's not.

You can use almost any sequence of characters
as a delimiter for the JavaScript code.
The same character sequence must be used
for the start and the end delimiter.
Exceptions are the pairs (), [], {} and <>.