support only a single is_mail function in validator_api
Hard-code options in the package body of validator_api.
Remove the implementation variant based on dbms_mle.
This way the function becomes a good example how
a subset of a JS library can be provided within the
Oracle Database. There is no need to provide all options
in the PL/SQL function.